Internal code refactor; remove top-level internal functions
